client/src/requirements.txt
Flask==2.0.2
Flask-Cors==3.0.10
httplib2==0.20.2
psycopg2==2.9.5
SQLAlchemy==1.4.17
urllib3==1.26.8
dataclass-type-validator==0.1.2
requests==2.28.1
pycryptodome==3.17